Rewrote functions and called them and changed error messages
