+++ src/sourmash/cli/utils.py
@@ -2,6 +2,8 @@ from glob import glob
 import os
 import argparse
 from sourmash.logging import notify
+from sourmash.sourmash_args import check_scaled_bounds
+from sourmash.sourmash_args import check_num_bounds
 
 
 def add_moltype_args(parser):
@@ -96,20 +98,20 @@ def command_list(dirpath):
     return sorted(basenames)
 
 
-def check_scaled_bounds(arg):
-    actual_min_val = 0
-    min_val = 100
-    max_val = 1e6
+# def check_scaled_bounds(arg):
+#     actual_min_val = 0
+#     min_val = 100
+#     max_val = 1e6
 
-    f = float(arg)
+#     f = float(arg)
 
-    if f < actual_min_val:
-        raise argparse.ArgumentTypeError(f"ERROR: --scaled value must be positive")
-    if f < min_val:
-        notify('WARNING: --scaled value should be >= 100. Continuing anyway.')
-    if f > max_val:
-        notify('WARNING: --scaled value should be <= 1e6. Continuing anyway.')
-    return f
+#     if f < actual_min_val:
+#         raise argparse.ArgumentTypeError(f"ERROR: --scaled value must be positive")
+#     if f < min_val:
+#         notify('WARNING: --scaled value should be >= 100. Continuing anyway.')
+#     if f > max_val:
+#         notify('WARNING: --scaled value should be <= 1e6. Continuing anyway.')
+#     return f
 
 
 def add_scaled_arg(parser, default=None):
@@ -119,20 +121,20 @@ def add_scaled_arg(parser, default=None):
     )
 
 
-def check_num_bounds(arg):
-    actual_min_val = 0
-    min_val = 50
-    max_val = 50000
+# def check_num_bounds(arg):
+#     actual_min_val = 0
+#     min_val = 50
+#     max_val = 50000
 
-    f = int(arg)
+#     f = int(arg)
 
-    if f < actual_min_val:
-        raise argparse.ArgumentTypeError(f"ERROR: --num-hashes value must be positive")
-    if f < min_val:
-        notify('WARNING: --num-hashes value should be >= 50. Continuing anyway.')
-    if f > max_val:
-        notify('WARNING: --num-hashes value should be <= 50000. Continuing anyway.')
-    return f
+#     if f < actual_min_val:
+#         raise argparse.ArgumentTypeError(f"ERROR: --num-hashes value must be positive")
+#     if f < min_val:
+#         notify('WARNING: --num-hashes value should be >= 50. Continuing anyway.')
+#     if f > max_val:
+#         notify('WARNING: --num-hashes value should be <= 50000. Continuing anyway.')
+#     return f
 
 
 def add_num_arg(parser, default=0):
